Moved captcha code from survey model into it's own trait
